pyproject.toml
[tool.poetry]
name = "ecommerce_be"
version = "0.1.0"
description = ""
authors = ["siegerts"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.109.2"
uvicorn = "0.27.0.post1"
sqlmodel = "0.0.14"
psycopg = "3.1.18"
python-decouple = "3.8"
annotated-types = "0.6.0"
anyio = "4.2.0"
certifi = "2025.6.15"
charset-normalizer = "3.4.2"
click = "8.1.7"
colorama = "0.4.6"
dependency-injector = "4.48.1"
greenlet = "3.0.3"
h11 = "0.14.0"
httptools = "0.6.1"
idna = "3.6"
mmh3 = "4.1.0"
psycopg-binary = "3.1.18"
psycopg2 = "2.9.10"
pydantic = {version = "2.6.1",  extras = ["email"]}
pydantic-core = "2.16.2"
python-dotenv = "1.0.1"
pyyaml = "6.0.1"
redis = "6.2.0"
requests = "2.32.3"
rstream = "0.30.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.25"
starlette = "0.36.3"
typing-extensions = "4.14.0"
tzdata = "2023.4"
urllib3 = "2.5.0"
watchfiles = "0.21.0"
websockets = "12.0"
PyJWT = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
